requirements.txt
Pillow
numpy
pyperclip